+++ editor.planx.uk/src/@planx/components/shared/Schema/InputFields/DateFieldInput.tsx
@@ -1,0 +1,32 @@
+import Box from "@mui/material/Box";
+import { paddedDate } from "@planx/components/DateInput/model";
+import type { DateField } from "@planx/components/shared/Schema/model";
+import React from "react";
+import InputLabel from "ui/public/InputLabel";
+import DateInput from "ui/shared/DateInput";
+
+import { getFieldProps, Props } from ".";
+
+export const DateFieldInput: React.FC<Props<DateField>> = (props) => {
+  const { data, formik } = props;
+  const { id, errorMessage, name, value } = getFieldProps(props);
+
+  if (typeof value !== "string")
+    throw Error("'value' prop for DateFieldInput must be of type string");
+
+  return (
+    <InputLabel label={data.title} htmlFor={id}>
+      <Box sx={{ display: "flex", alignItems: "baseline" }}>
+        <DateInput
+          value={value}
+          bordered
+          onChange={(newDate: string, eventType: string) => {
+            formik.setFieldValue(name, paddedDate(newDate, eventType));
+          }}
+          error={errorMessage}
+          id={id}
+        />
+      </Box>
+    </InputLabel>
+  );
+};
